Revert "评价课次接口字段json序列化兼容"
This reverts commit 6ae04a8ac556b259208b0867bd761900820421ad.

+++ eams-cpp/eams-c1-Home/controller/class/classcontroller.cpp
@@ -21,7 +21,7 @@
 #include "SimpleDateTimeFormat.h"
 #include "../../service/class/classService.h"
 
-classInfoJsonVO::Wrapper classController::execQueryClassInfo(const UInt64 &class_id) {
+classInfoJsonVO::Wrapper classController::execQueryClassInfo(const UInt32 &class_id) {
 	auto res = ClassService().getById(class_id);
 	if (res == nullptr) {
 		auto jvo = classInfoJsonVO::createShared();
@@ -41,7 +41,7 @@ classPageJsonVO::Wrapper classController::execQueryClassPage(const classQuery::W
 	return jvo;
 }
 
-studentListJsonVO::Wrapper classController::execQueryStudentList(const UInt64& class_id) {
+studentListJsonVO::Wrapper classController::execQueryStudentList(const UInt32& class_id) {
 	auto res = StudentService().listByClassId(class_id);
 	auto jvo = studentListJsonVO::createShared();
 	jvo->success(res);
