+++ eams-cpp/eams-c1-Home/domain/query/attendance-records/attendancequery.h
@@ -32,7 +32,7 @@ class attendance_recordsQuery : public PageQuery
 {
 	DTO_INIT(attendance_recordsQuery, PageQuery);
 	// 学生id
-	API_DTO_FIELD_REQUIRE(UInt64, student_id, ZH_WORDS_GETTER("attendance_records.student_id"), true);
+	API_DTO_FIELD_REQUIRE(UInt32, student_id, ZH_WORDS_GETTER("attendance_records.student_id"), true);
 	// 课程id
 	//API_DTO_FIELD_DEFAULT(UInt32, course_id, ZH_WORDS_GETTER("attendance_records.course_id"));
 	// 上课日期
